app/musicgen-app/backend/requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
httpx
python-jose[cryptography]
redis
stripe
asyncpg
fastapi-limiter
